requirements.txt
chromadb
azure-core
azure-search-documents
python-docx
openai
Flask